requirements.txt
beautifulsoup4==4.12.2
boto3==1.12.41
botocore==1.15.49
Brotli==1.0.7
bsdiff4==1.1.9
certifi==2023.11.17
charset-normalizer==3.3.2
Click==7.0
colorama==0.4.6
docopt==0.6.2
docutils==0.15.2
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
idna==3.6
iniconfig==2.0.0
interchange==2021.0.4
jmespath==0.10.0
loguru==0.7.2
monotonic==1.6
neobolt==1.7.17
neotime==1.7.4
packaging==23.2
pansi==2020.7.3
pathlib==1.0.1
pluggy==1.3.0
progressbar==2.5
prompt-toolkit==2.0.10
protobuf==3.11.3
py2neo==2021.2.4
pycryptodome==3.9.7
pymongo==3.10.1
pytest==7.4.3
python-dateutil==2.8.2
python-magic==0.4.15
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.3.7
six==1.16.0
soupsieve==2.5
urllib3==1.24.3
wcwidth==0.2.12
Pygments>=2.3.1
androguard==3.3.5
oscrypto==1.3.0
pycryptodomex
google3
